move test utils under __tests__ and use testPathIgnorePatterns
